# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "astrea-sda-api",
    "description": "A high-performance Rust API for Space Domain Awareness (SDA), satellite tracking, SGP4 propagation, and visibility prediction",
    "contact": {
      "name": "Michael Wedd"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "OpenAPI endpoint handler to issue an RS256 JWT authentication token for testing or client auth.",
        "operationId": "loginHandler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "RS256 JWT Token issued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthResponse": {
        "type": "object",
        "description": "Response returned upon successful RS256 JWT token generation.",
        "required": [
          "token",
          "tokenType",
          "expiresIn",
          "claims"
        ],
        "properties": {
          "claims": {
            "$ref": "#/components/schemas/Claims"
          },
          "expiresIn": {
            "type": "integer",
            "description": "Token lifetime in seconds (e.g. 86400)",
            "minimum": 0
          },
          "token": {
            "type": "string",
            "description": "Encoded RS256 JWT Bearer token"
          },
          "tokenType": {
            "type": "string",
            "description": "Token type (always \"Bearer\")"
          }
        }
      },
      "Claims": {
        "type": "object",
        "description": "JWT Claims payload structure containing standard OIDC and OAuth 2.0 claim fields.",
        "required": [
          "sub",
          "exp",
          "iat",
          "role"
        ],
        "properties": {
          "aud": {
            "type": "string",
            "description": "Token audience (target API audience)",
            "nullable": true
          },
          "exp": {
            "type": "integer",
            "description": "Expiration timestamp in seconds since Unix epoch",
            "minimum": 0
          },
          "iat": {
            "type": "integer",
            "description": "Issued-at timestamp in seconds since Unix epoch",
            "minimum": 0
          },
          "iss": {
            "type": "string",
            "description": "Token issuer (OIDC issuer URI or auth server ID)",
            "nullable": true
          },
          "role": {
            "type": "string",
            "description": "User role (\"viewer\", \"editor\", or \"admin\")"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of secondary roles",
            "nullable": true
          },
          "scope": {
            "type": "string",
            "description": "Space-separated OAuth 2.0 scopes string (e.g. \"read:satellites write:satellites admin:satellites\")",
            "nullable": true
          },
          "sub": {
            "type": "string",
            "description": "Subject (User ID, username, or client service ID)"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "status"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "description": "Request body for JWT authentication login/token generation endpoint.",
        "required": [
          "username"
        ],
        "properties": {
          "role": {
            "type": "string",
            "description": "Requested role (\"viewer\", \"editor\", or \"admin\", defaults to \"viewer\")",
            "nullable": true
          },
          "scope": {
            "type": "string",
            "description": "Optional OAuth 2.0 scopes string (e.g. \"read:satellites write:satellites\")",
            "nullable": true
          },
          "username": {
            "type": "string",
            "description": "Username or service subject identifier"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}